5genUntiに対応

+++ PWTSeed.cs
@@ -8,16 +8,16 @@ namespace PWT_RNG
 {
     internal class PWTSeed
     {
-        private uint Seed;
-        private uint Count;
-        uint[] stateVector = new uint[624];
+        private ulong Seed;
+        private ulong Count;
+        ulong[] stateVector = new ulong[624];
         protected const int N = 624;
         protected const int M = 397;
-        protected const uint MATRIX_A = 0x9908b0df;
-        protected const uint UPPER_MASK = 0x80000000;
-        protected const uint LOWER_MASK = 0x7fffffff;
+        protected const ulong MATRIX_A = 0x9908b0df;
+        protected const ulong UPPER_MASK = 0x80000000;
+        protected const ulong LOWER_MASK = 0x7fffffff;
 
-        public PWTSeed(uint Seed, uint Count)
+        public PWTSeed(ulong Seed, ulong Count)
         {
             this.Seed = Seed;
             this.Count = Count;
@@ -50,9 +50,8 @@ namespace PWT_RNG
         }
         public ulong PWTRNG()
         {
-
-            uint val1, val2, val3;
-            uint temp1, temp2, temp3, temp4, temp5, temp6;
+            ulong val1, val2, val3;
+            ulong temp1, temp2, temp3, temp4, temp5, temp6;
             ulong PWT, PWT1, PWT2, PWTtemp;
 
             val1 = stateVector[Count + 1];
@@ -70,7 +69,6 @@ namespace PWT_RNG
 
             PWT = (PWT1 << 0x20) + PWT2;    //PWT乱数(仮)
             return PWT;
-
         }
     } 
 }
